# By Aleksei Kanash
# Via Aleksei Kanash
* commit '0cc44a74145a5da727681e3b90032da0a369cf30 [formerly 24cf36f4c2cd5786a8de6c8eb59c1e495846e95a]':
  [DF-7335] Color uses C-style array to keep compatibility to external code.
  [DF-7335] Codestyle fix.
  [DF-7335] NMaterial properties is Vectors now.
  [DF-7335] removed COUNT_OF after development merge.
  [DF-7335] Review fixes, correct isdigit usage.
  [DF-7335] Win 10 warnings fix.
  [DF-7335] std::array removed from Vectors.
  [DF-7335] Review fix. Compilation fix.
  [DF-7335] Review fixes.
  [DF-7335] Vegetation system has no COUNT_OF usage. Vectors usage refactor.
  [DF-7335] Vegetation Geometry arrays uses std::array and std::vector.
  [DF-7335] Removed offsetof as non compilable on Mac. Count_Of method usage removed almost everywhere.
  [DF-7335] NMaterial don't uses COUNT_OF.
  [DF-7335] Review fixes, more  nullptr instead of NULL in changed files.
  [DF-7335] Some PVS warnings fixed.
  [DF-7335] Warnings fix for windows x64


Former-commit-id: 21f21f9aaa9d2ef31a240df8d8be8fb0462874f7

+++ Sources/Internal/Utils/UtilsWin32.cpp
@@ -50,7 +50,7 @@ String GenerateGUID()
 
     //get string representation of GUID
     Array<OLECHAR, 64> guidString {};
-    ::StringFromGUID2(guid, guidString.data(), guidString.size());
+    ::StringFromGUID2(guid, guidString.data(), static_cast<int>(guidString.size()));
     
     //convert to normal string
     USES_CONVERSION;
